Alteração para atender ao desafio
Altera para variar diminuindo o tempo necessário para clique, score aumenta de acordo com a dificuldade e velocidade de clique.

+++ src/miniGame/model/score/Score.java
@@ -2,12 +2,12 @@ package miniGame.model.score;
 
 public class Score implements Comparable<Score>{
     private String player;
-    private int score;
+    private float score;
 
     public Score() {
 
     }
-    public Score(String gamer, int score) {
+    public Score(String gamer, float score) {
         this.player = gamer;
         this.score = score;
     }
